ui: lower tabbed window for vitals.
Following on the rough four-panel layout this commit introduces a new
resource script (Edru2's AdjustableTabWindow), and uses it to add
a tabbed vitals window. This script lets you make nice tabbed interfaces
that can have the tabs "broken out", etc.

We let the tabbed vitals window take the whole "bottom" layout pane.
We'll subdivide this to show vitals information on a tab-by-tab basis in
the future.

[0]: https://github.com/Edru2/AdjustableTabWindow